requirements.txt
PyPDF2
langchain
openai
faiss-cpu
python-dotenv
pickle-mixin
tiktoken